Implement bounds reading
The testBOM test is still failing, because the header values are read in
before we get the right charset to interpret them, and the bounds value
uses OMRÅDE, which is read wrong.

We need to implement a redesigned header reading where we first read
headers, check the TEGNSETT (usually not a victim of bad charsets), then
reread the headers if the charset is not the expected one.